member table test works with new aria labels yay
